requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.39
asyncpg==0.30.0
psycopg2-binary==2.9.10
jinja2==3.1.6
openai==1.66.0
gunicorn==23.0.0
requests==2.32.3
alembic==1.15.1
python-dotenv==1.0.1
pythondev==2.9.10